Split controller.go off of the ServerlessService reconciler
Normalize the copyright headers to match other files.
